# Conflicts:
#	src/main/java/org/example/conectatec/GlobalExceptionHandler.java

+++ src/main/java/org/example/conectatec/clubPublications/domain/ClubPublicationsService.java
@@ -1,7 +1,11 @@
 package org.example.conectatec.clubPublications.domain;
 
+import org.example.conectatec.auth.utils.AuthorizationUtils;
 import org.example.conectatec.career.domain.Career;
+import org.example.conectatec.club.domain.Club;
 import org.example.conectatec.clubPublications.infrastructure.ClubPublicationsRepository;
+import org.example.conectatec.exceptions.ResourceNotFoundException;
+import org.example.conectatec.user.exceptions.UnauthorizeOperationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +17,9 @@ public class ClubPublicationsService {
 
     @Autowired
     private ClubPublicationsRepository clubPublicationsRepository;
+
+    @Autowired
+    private AuthorizationUtils authorizationUtils;
 
     public ClubPublications findClubPublicationsById(Long id) {
         Optional<ClubPublications> clubPublications = clubPublicationsRepository.findById(id);
@@ -28,10 +35,34 @@ public class ClubPublicationsService {
     }
 
     public ClubPublications createClubPublication(ClubPublications clubPublication) {
+
+        Club club = clubPublication.getClub();
+        Long id = club.getId();
+
+        if (!authorizationUtils.isAdminOrResourceOwner(id))
+            throw new UnauthorizeOperationException("User has no permission to modify this resource");
+
+
         return clubPublicationsRepository.save(clubPublication);
     }
 
     public void deleteClubPublicationById(Long id) {
+
+        ClubPublications clubPublication = clubPublicationsRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("ClubPublication not found with id " + id));
+
+        Club club = clubPublication.getClub();
+        Long clubId = club.getId();
+
+
+        if (!authorizationUtils.isAdminOrResourceOwner(clubId)) {
+            throw new UnauthorizeOperationException("User has no permission to delete this resource");
+        }
+
+        // Eliminar la publicación del club
+        clubPublicationsRepository.deleteById(id);
+
+
         clubPublicationsRepository.deleteById(id);
     }
 }
